stuck on adding the trip model to the view, so using a placeholder with the most update ui for dashboard now
